Fixed the functions to use the types created in sugarfunge-api-types
